+++ src/sparqlmodel/stores/__init__.py
@@ -4,6 +4,7 @@ from sparqlmodel.stores.async_http import AsyncHttpStore
 from sparqlmodel.stores.async_memory import AsyncMemoryStore
 from sparqlmodel.stores.base import Store
 from sparqlmodel.stores.http import HttpStore
+from sparqlmodel.stores.http_common import QueryMethod
 from sparqlmodel.stores.memory import MemoryStore
 
 __all__ = [
@@ -11,5 +12,6 @@ __all__ = [
     "AsyncMemoryStore",
     "HttpStore",
     "MemoryStore",
+    "QueryMethod",
     "Store",
 ]